+++ grisked-profile/src/models/label.rs
@@ -2,13 +2,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Clone, Debug)]
 pub struct Label {
-    pub id: u16,
+    pub id: usize,
     pub name: String,
     pub color: [f32; 3],
 }
 
 impl Label {
-    pub fn new(id: u16, name: String, color: [f32; 3]) -> Self {
+    pub fn new(id: usize, name: String, color: [f32; 3]) -> Self {
         for element in color {
             debug_assert!(
                 (0.0..=1.0).contains(&element),
